# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: DiAkoOi API
  version: 0.2.0
  description: >-
    Feedback and word-bank delivery for DiAkoOi.


    **Nothing here is required to start or finish a game.** The app is
    offline-first, ships a complete word bank, and treats the server copy as
    an update rather than a dependency (CLAUDE.md §Hard rules). Every endpoint
    is allowed to be unreachable, and the app must degrade silently when it is
    (docs/12-HOSTING.md §2c).


    **There is no telemetry endpoint, deliberately.** v1 collects nothing —
    see docs/adr/0015-no-telemetry-in-v1.md. A documented endpoint that
    returns 501 is a contract to maintain for nothing, so it is absent rather
    than stubbed.


    **No endpoint accepts a photograph of a person.** Selfies live in device
    memory only and never leave it (01-DESIGN.md §4b). Feedback screenshots
    are a different consent model entirely — the user chose to send one — and
    are constrained by media type and size below.
  license:
    name: See LICENSE in the repository
servers:
  - url: http://localhost:3000
    description: Local development
  # The public base URL rotates on every cloudflared restart and is therefore
  # NEVER baked into a client. The app resolves it at runtime from
  # endpoint.json (docs/12-HOSTING.md §2b).
  - url: https://{tunnel}.trycloudflare.com
    description: Beta, behind a Cloudflare Quick Tunnel
    variables:
      tunnel:
        default: example
        description: >-
...
paths:
...
  /v1/word-banks/{topic}:
    get:
      operationId: getWordBank
      tags: [word-banks]
      summary: Fetch one topic's words and authored clues
      parameters:
        - name: topic
          in: path
          required: true
          description: Topic id, as listed by `GET /v1/word-banks`.
          schema:
            $ref: '#/components/schemas/TopicId'
      responses:
        '200':
          description: The topic's words.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordBankResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
components:
...
  responses:
    ValidationError:
      description: The request body or parameters failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: >-
        The path exists but not for this method. Distinguished from 404
        deliberately: a client hitting the wrong verb should not be told the
        endpoint is missing.
      headers:
        Allow:
          description: Methods this path does accept.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: >-
        Rate limit exceeded. Every public endpoint is limited; a residential
        host has no capacity to absorb abuse (docs/12-HOSTING.md §4).
      headers:
        Retry-After:
          description: Seconds until the limit resets.
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    DatabaseUnavailable:
      description: >-
        The database is unreachable. The app treats this exactly like being
        offline — bundled content, no error dialog (docs/12-HOSTING.md §2c).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unhandled server fault. Never contains internal detail.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnexpectedError:
      description: Any other error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      additionalProperties: false
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
          description: Correlation id echoed in logs.
          minLength: 1
        timestamp:
          type: string
          format: date-time
    TopicId:
      type: string
      description: Stable topic identifier, lowercase.
      pattern: '^[a-z][a-z0-9_]{1,31}$'
...
    ClueSet:
      type: object
      additionalProperties: false
      required: [tight, standard, loose]
      description: >-
        Three authored tiers. Nothing here is algorithmically derived —
        derivation collapses at a table and produces uneven difficulty with no
        way to tune it (01-DESIGN.md §14).
      properties:
        tight:
          type: string
          minLength: 1
          maxLength: 200
        standard:
          type: string
          minLength: 1
          maxLength: 200
        loose:
          type: string
          minLength: 1
          maxLength: 200
    WordBankEntry:
      type: object
      additionalProperties: false
      required: [topicId, word, clues]
      properties:
        topicId:
          $ref: '#/components/schemas/TopicId'
        word:
          type: string
          minLength: 1
          maxLength: 80
        clues:
          $ref: '#/components/schemas/ClueSet'
        difficultyRating:
          type: integer
          minimum: 1
          maximum: 5
        region:
          type: string
          enum: [national, luzon, visayas, mindanao]
    WordBankData:
      type: object
      additionalProperties: false
      required: [topicId, contentVersion, words]
      properties:
        topicId:
          $ref: '#/components/schemas/TopicId'
        contentVersion:
          type: string
          minLength: 1
          maxLength: 64
        words:
          type: array
          items:
            $ref: '#/components/schemas/WordBankEntry'
    WordBankResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/WordBankData'
        meta:
          $ref: '#/components/schemas/Meta'

    # Documenting the error body is the cheapest way to avoid the most common
    # contract violation (docs/06-TESTING-STRATEGY.md §6).
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable identifier.
              minLength: 1
            message:
              type: string
              description: Human-readable explanation. Never contains secrets.
              minLength: 1
            details:
              type: array
              description: >-
                Field-level validation problems, when the code is
                VALIDATION_FAILED. Absent otherwise.
              items:
                type: object
                additionalProperties: false
                required: [path, message]
                properties:
                  path:
                    type: string
                    minLength: 1
                  message:
                    type: string
                    minLength: 1